Move FormattedOutput implementation of Variable
